split ast types
